User Directions

1. Open Cloud

You can try TheNumberLine.jl in the cloud without having to install anything.

2. Private Computer

If instead you want to use TheNumberLine.jl then proceed as follows.

Download TheNumberLine.jl

Open a terminal window go to the TheNumberLine.jl/src/ folder

Start Pluto.jl e.g. via julia -e "using Pluto; Pluto.run()"

In web browser window, select and open standard_number_line.jl

Tip

Some tips for new users are provided below.

To use TheNumberLine.jl on your own computer, you'll need to download Julia itself beforehand.

to close the session you want to go back to the terminal window and Press Ctrl+C

Main Functions

TheNumberLine.NumberLinePlot — Function

NumberLinePlot(y; path=tempdir(), filename="tmp.png")

Display the number line sequence of operations, defined by vector y, as a graph.

TheNumberLine.NumberLineExpression — Function

NumberLineExpression(x)

Display the number line sequence of operations, defined by vector x, as text.

TheNumberLine.make_hist — Function

make_hist(;max_value=5.0,max_count=5,aspect_ratio=2.0)

fig=make_hist()

TheNumberLine.save_hist — Function

save_hist(fig::Makie.Figure;file=joinpath(tempdir(),"histogram_template.png"))

save_hist(make_hist())

Various

1. Use a Slider

In the first code cell of standard_number_line.jl, where values are entered, you can use a list of a slider or multiple choice instead.

To do this, replace the line that says

` Type value : ` $(@bind b html"<input type=text>")

with

`Select value of x from the list : ` $(@bind b Select(["4","3","2","1","0","-1","-2","-3","-4"],default="0"))

or

`Choose value of x using slider : ` $(@bind b aSlider(-10:10; default=0))

2. Integers Only

To restrict the number line to using integers (as opposed to reals), try changing TestType in standard_number_line.jl.

To do this, replace the line that says

TestType=Number; NumberLineType=Float64; "Number LineType = $(NumberLineType)"

with

TestType=Int; NumberLineType=Int; "NumberLineType = $(NumberLineType)"