User Directions

1. Open Cloud

You can try TheNumberLine.jl in the cloud without having to install anything.

2. Private Computer

If instead you want to use TheNumberLine.jl then proceed as follows:

Download TheNumberLine.jl

Start Pluto.jl in a terminal window

julia -e 'using Pluto; Pluto.run(notebook="TheNumberLine.jl/src/standard_number_line.jl")'

The result should look as follows:

standard_number_line.html

standard_number_line.jl

Tip

To use TheNumberLine.jl on your own computer, you'll need to download Julia itself beforehand.

to close the session you want to go back to the terminal window and Press Ctrl+C

Main Functions

TheNumberLine.NumberLinePlot — Function

NumberLinePlot(y; path=tempdir(), filename="tmp.png")

Display the number line sequence of operations, defined by vector y, as a graph.

TheNumberLine.NumberLineExpression — Function

NumberLineExpression(x)

Display the number line sequence of operations, defined by vector x, as text.

TheNumberLine.make_hist — Function

make_hist(;max_value=5.0,max_count=5,aspect_ratio=2.0)

fig=make_hist()

TheNumberLine.save_hist — Function

save_hist(fig::Makie.Figure;file=joinpath(tempdir(),"histogram_template.png"))

save_hist(make_hist())

Various

1. Use a Slider

In the first code cell of standard_number_line.jl, where values are entered, you can use a list of a slider or multiple choice instead.

To do this, replace the line that says

` Type value : ` $(@bind b html"<input type=text>")

with

`Select value of x from the list : ` $(@bind b Select(["4","3","2","1","0","-1","-2","-3","-4"],default="0"))

or

`Choose value of x using slider : ` $(@bind b aSlider(-10:10; default=0))

2. Integers Only

To restrict the number line to using integers (as opposed to reals), try changing TestType in standard_number_line.jl.

To do this, replace the line that says

TestType=Number; NumberLineType=Float64; "Number LineType = $(NumberLineType)"

with

TestType=Int; NumberLineType=Int; "NumberLineType = $(NumberLineType)"

School Day Table

table_school_days.html

table_school_days.jl

TheNumberLine.table_school_days — Function

table_school_days(fil=joinpath(tempname,".html"))

Generate waterfall calendar, in html or csv format.

Use example:

using TheNumberLine

day1="2024-08-27"
dayN="2025-06-18"
bloctimes=["8:00am","8:53am","8:56am","9:49am","9:52am","10:45am",
    "11:11am","12:26pm","12:29pm","1:23pm","1:26pm","2:20pm"]
holidays=["2024-08-30","2024-09-02","2024-09-03","2024-10-14",
    "2024-11-05","2024-11-11","2024-11-28","2024-11-29","2024-12-23",
    "2024-12-24","2024-12-25","2024-12-26","2024-12-27","2024-12-28",
    "2024-12-29","2024-12-30", "2024-12-31", "2025-01-01","2025-01-20",
    "2025-02-17","2025-02-18","2025-02-19","2025-02-20","2025-02-21",
    "2025-04-21","2025-04-22","2025-04-23","2025-04-24","2025-04-25","2025-05-26"]
halfdays=["2024-09-18","2024-10-16","2024-11-07","2024-11-13",
    "2024-11-27","2024-12-18","2024-12-20","2025-01-15","2025-02-06",
    "2025-02-12","2025-03-19","2025-04-09","2025-05-14","2025-06-11"]
 
table_school_days(file="my_school_days.html",
    day1=day1,dayN=dayN,bloctimes=bloctimes,
    holidays=holidays,halfdays=halfdays)

or for csv output just change the file name extension :

table_school_days(file="my_school_days.csv",
    day1=day1,dayN=dayN,bloctimes=bloctimes,
    holidays=holidays,halfdays=halfdays)